fit into selections logic
